move all map tests to one file
