+++ backend/dashboard/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from django.http import JsonResponse
 from rest_framework.decorators import api_view
-from .use_cases import DashboardUseCases
+from .use_case import DashboardUseCases
 from random import randint
 
 class DashboardView:
